Moved definitions from LocomoCore/Source/Public/Distance/AtypicalDistances.h into LocomoCore/Source/Public/AtypicalDistances.h and removed duplicate class
